+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "klinecharts",
-  "version": "8.3.6",
+  "version": "8.5.0",
   "description": "Lightweight k-line chart built with html5 canvas",
   "main": "index.js",
   "types": "./types/index.d.ts",
