--- types/Pane.d.ts
@@ -1,5 +1,0 @@
-export declare interface PaneOptions {
-  id?: string;
-  height?: number;
-  dragEnabled?: boolean;
-}
