--- types/Tag.d.ts
@@ -1,10 +1,0 @@
-import { Point } from './Common';
-
-export declare interface Tag {
-  id: string;
-  point: Point;
-  text?: string | number;
-  mark?: string;
-  coordinate?: number;
-  styles?: any;
-}
